+++ test/scheduler.ts
@@ -8,6 +8,7 @@ import {
   createApolloStore,
 } from '../src/store';
 import mockNetworkInterface from './mocks/mockNetworkInterface';
+import { NetworkStatus } from '../src/queries/store';
 import gql from 'graphql-tag';
 
 describe('QueryScheduler', () => {
@@ -76,7 +77,7 @@ describe('QueryScheduler', () => {
       queryManager,
     });
     let timesFired = 0;
-    const queryId = scheduler.startPollingQuery(queryOptions, 'fake-id', true, (queryStoreValue) => {
+    const queryId = scheduler.startPollingQuery(queryOptions, 'fake-id', (queryStoreValue) => {
       timesFired += 1;
     });
     setTimeout(() => {
@@ -106,7 +107,9 @@ describe('QueryScheduler', () => {
     };
     const networkInterface = mockNetworkInterface(
       {
-        request: queryOptions,
+        request: {
+          query: queryOptions.query,
+        },
         result: { data },
       }
     );
@@ -119,9 +122,11 @@ describe('QueryScheduler', () => {
       queryManager,
     });
     let timesFired = 0;
-    let queryId = scheduler.startPollingQuery(queryOptions, 'fake-id', true, (queryStoreValue) => {
-      timesFired += 1;
-      scheduler.stopPollingQuery(queryId);
+    let queryId = scheduler.startPollingQuery(queryOptions, 'fake-id', (queryStoreValue) => {
+      if (queryStoreValue.networkStatus !== NetworkStatus.poll) {
+        timesFired += 1;
+        scheduler.stopPollingQuery(queryId);
+      }
     });
 
     setTimeout(() => {
@@ -257,55 +262,9 @@ describe('QueryScheduler', () => {
     });
   });
 
-  it('should keep track of in flight queries', (done) => {
-    const query = gql`
-      query {
-        fortuneCookie
-      }`;
-    const data = {
-      'fortuneCookie': 'lol',
-    };
-    const queryOptions = {
-      query,
-      pollInterval: 70,
-      forceFetch: true,
-    };
-    const networkInterface = mockNetworkInterface(
-      {
-        request: queryOptions,
-        result: { data },
-        delay: 20000, //i.e. should never return
-      },
-      {
-        request: queryOptions,
-        result: { data },
-        delay: 20000,
-      }
-    );
-    const queryManager = new QueryManager({
-      networkInterface,
-      store: createApolloStore(),
-      reduxRootSelector: defaultReduxRootSelector,
-    });
-    const scheduler = new QueryScheduler({
-      queryManager,
-    });
-    const observer = scheduler.registerPollingQuery(queryOptions);
-    const subscription = observer.subscribe({});
-
-    // as soon as we register a query, there should be an addition to the query map.
-    assert.equal(Object.keys(scheduler.inFlightQueries).length, 1);
-    setTimeout(() => {
-      assert.equal(Object.keys(scheduler.inFlightQueries).length, 1);
-      assert.deepEqual(scheduler.inFlightQueries[0], queryOptions);
-      subscription.unsubscribe();
-      done();
-    }, 100);
-  });
-
   it('should not fire another query if one with the same id is in flight', (done) => {
     const query = gql`
-      query {
+      query B {
         fortuneCookie
       }`;
     const data = {
@@ -333,7 +292,6 @@ describe('QueryScheduler', () => {
     const observer = scheduler.registerPollingQuery(queryOptions);
     const subscription = observer.subscribe({});
     setTimeout(() => {
-      assert.equal(Object.keys(scheduler.inFlightQueries).length, 1);
       subscription.unsubscribe();
       done();
     }, 100);
@@ -527,7 +485,7 @@ describe('QueryScheduler', () => {
       queryManager,
     });
     let timesFired = 0;
-    const queryId = scheduler.startPollingQuery(queryOptions, 'fake-id', true, (queryStoreValue) => {
+    const queryId = scheduler.startPollingQuery(queryOptions, 'fake-id', (queryStoreValue) => {
       timesFired += 1;
     });
     setTimeout(() => {
@@ -571,11 +529,11 @@ describe('QueryScheduler', () => {
       queryManager,
     });
     let timesFired = 0;
-    let queryId = scheduler.startPollingQuery(queryOptions, 'fake-id', true, (queryStoreValue) => {
+    let queryId = scheduler.startPollingQuery(queryOptions, 'fake-id', (queryStoreValue) => {
       scheduler.stopPollingQuery(queryId);
     });
     setTimeout(() => {
-      let queryId2 = scheduler.startPollingQuery(queryOptions, 'fake-id2', true, (queryStoreValue) => {
+      let queryId2 = scheduler.startPollingQuery(queryOptions, 'fake-id2', (queryStoreValue) => {
         timesFired += 1;
       });
       assert.equal(scheduler.intervalQueries[20].length, 1);